create notification from date

+++ lib/models/my_notification.dart
@@ -4,7 +4,7 @@ part 'my_notification.g.dart';
 @HiveType(typeId: 88)
 class MyNotification {
   @HiveField(0)
-  final BigInt? id;
+  final int? id;
   @HiveField(1)
   final DateTime? dateStart;
   @HiveField(2)
